refactor: deepen playthrough modules
