+++ src/components/team/teamMemberSelectionDomain.ts
@@ -1,0 +1,6 @@
+import type { PokemonOptionType } from "@/loaders/pokemon";
+
+export const getTeamSelectionNickname = (
+  headPokemon: PokemonOptionType | null | undefined,
+  bodyPokemon: PokemonOptionType | null | undefined,
+) => headPokemon?.nickname ?? bodyPokemon?.nickname ?? "";
