+++ src/stores/playthroughs/encounters/dragDrop.ts
@@ -62,6 +62,52 @@ export const clearEncounterFromLocation = async (
   }
 };
 
+export const relocateEncounterSlot = async ({
+  sourceLocationId,
+  sourceField,
+  targetLocationId,
+  targetField,
+}: {
+  sourceLocationId: string;
+  sourceField: "head" | "body";
+  targetLocationId: string;
+  targetField: "head" | "body";
+}) => {
+  if (sourceLocationId === targetLocationId && sourceField === targetField) {
+    return;
+  }
+
+  const activePlaythrough = ensureActivePlaythroughWithEncounters();
+  if (!activePlaythrough) {
+    return;
+  }
+
+  const sourceEncounter = activePlaythrough.encounters[sourceLocationId];
+  const pokemon = sourceEncounter?.[sourceField];
+  if (!pokemon) {
+    return;
+  }
+
+  const targetEncounter = activePlaythrough.encounters[targetLocationId];
+  if (targetEncounter?.[targetField]) {
+    await swapEncounters(
+      sourceLocationId,
+      targetLocationId,
+      sourceField,
+      targetField,
+    );
+    return;
+  }
+
+  await moveEncounterAtomic(
+    sourceLocationId,
+    sourceField,
+    targetLocationId,
+    targetField,
+    pokemon,
+  );
+};
+
 // Move encounter atomically from source to destination (for drag and drop)
 export const moveEncounterAtomic = async (
   sourceLocationId: string,
