+++ src/stores/playthroughs/store.ts
@@ -1,6 +1,5 @@
 import { proxy, subscribe } from "valtio";
 import { devtools } from "valtio/utils";
-import { z } from "zod";
 import { getSharedEventProperties } from "@/lib/analytics/selectors";
 import {
   type SourceSurface,
@@ -11,7 +10,7 @@ import type { PokemonOptionType } from "@/loaders/pokemon";
 import { buildPokemonUidIndex } from "@/utils/encounter-utils";
 import { generatePrefixedId } from "@/utils/id";
 import { createDefaultPlaythrough } from "./defaultPlaythrough";
-import { migrateImportedPlaythroughData } from "./migrations";
+import { prepareImportedPlaythrough } from "./importPipeline";
 import {
   createDebouncedSaveAll,
   deletePlaythroughFromIndexedDB,
@@ -326,80 +325,15 @@ const getGameMode = (): GameMode => {
 };
 
 const importPlaythrough = async (importData: unknown): Promise<string> => {
-  try {
-    const { ImportedPlaythroughSchema } = await import("./types");
-    const migratedImportData = migrateImportedPlaythroughData(importData);
-    const validationResult =
-      ImportedPlaythroughSchema.safeParse(migratedImportData);
+  const newPlaythrough = prepareImportedPlaythrough(
+    importData,
+    playthroughsStore.playthroughs.map((p) => p.id),
+  );
 
-    if (!validationResult.success) {
-      throw validationResult.error;
-    }
+  playthroughsStore.playthroughs.push(newPlaythrough);
+  playthroughsStore.activePlaythroughId = newPlaythrough.id;
 
-    const validatedData = validationResult.data;
-
-    // Extract the playthrough data - the transform ensures proper typing
-    const importedPlaythrough = validatedData.playthrough;
-
-    // Check for ID conflicts and generate new ID if needed
-    const existingIds = new Set(
-      playthroughsStore.playthroughs.map((p) => p.id),
-    );
-    let finalId = importedPlaythrough.id;
-
-    if (existingIds.has(finalId)) {
-      // Generate a new unique ID with timestamp and crypto-secure random suffix
-      finalId = generatePrefixedId("playthrough");
-    }
-
-    // Create the new playthrough with migrated data
-    const newPlaythrough: Playthrough = {
-      id: finalId,
-      name: importedPlaythrough.name,
-      gameMode: importedPlaythrough.gameMode,
-      version: importedPlaythrough.version || "1.0.0",
-      createdAt: importedPlaythrough.createdAt,
-      updatedAt: Date.now(), // Update to current time
-      customLocations: importedPlaythrough.customLocations || [],
-      encounters: importedPlaythrough.encounters || {},
-      team: importedPlaythrough.team || {
-        members: [null, null, null, null, null, null],
-      }, // Fixed size 6 with null values
-    };
-
-    // Add to store
-    playthroughsStore.playthroughs.push(newPlaythrough);
-
-    // Set as active playthrough
-    playthroughsStore.activePlaythroughId = finalId;
-
-    return finalId;
-  } catch (error) {
-    console.error("Failed to import playthrough:", error);
-
-    // Handle Zod validation errors specifically
-    if (error && typeof error === "object" && "issues" in error) {
-      try {
-        const prettyError = z.prettifyError(error as z.ZodError);
-        throw new Error(`Validation failed:\n\n${prettyError}`);
-      } catch {
-        const zodError = error as z.ZodError;
-        if (zodError.issues && zodError.issues.length > 0) {
-          const errorDetails = zodError.issues
-            .map((issue: z.ZodIssue) => {
-              const path =
-                issue.path.length > 0 ? ` at ${issue.path.join(".")}` : "";
-              return `• ${issue.message}${path}`;
-            })
-            .join("\n");
-          throw new Error(`Validation failed:\n\n${errorDetails}`);
-        }
-        throw new Error("Data validation failed");
-      }
-    }
-
-    throw new Error("Invalid playthrough data format");
-  }
+  return newPlaythrough.id;
 };
 
 const isRandomizedModeEnabled = (): boolean => {
